Create a module for running functionality
Split the old main.py into a module and script
